Move expand_labels from measure to segmentation
